fix: avoid initially setting username if they already had one
We were keying off of the user.name to see if we should call the effect
to update their name. Now we set a ref that indicates if it was a user
initiated event that updated their name or if it was the initial load
of the page ... whether this is optimal or not I still do not know.